feat: add main layer, create adapter for routes, and add gorilla mux

+++ models/http.go
@@ -7,7 +7,7 @@ type HttpRequest struct {
 }
 
 type HttpResponse struct {
-	Code int16 `json:"code"`
+	Code int `json:"code"`
 	Body interface{} `json:"body"`
 }
